requirements.txt
PyMySQL==1.0.2
python-ldap==3.3.1
Django==3.2
django-auth-ldap==2.4.0
django-filter==2.4.0
django-simpleui==2021.4.3
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Markdown==3.3.3
Pillow==8.2.0